code review: user email prop 추가, 모달 컴포넌트 하나의 객체로 export 하도록 수정
